fix(iOS): Check for multiple screens while changing screen orientation (#2035)
## Description
This fixes two issues with scenes on iOS:

1. A casting issue, which allows non-window scenes on iOS, e.g.
CarPlay's CPTemplateApplicationScene
3. An assumed loading order - the UIWindowScene is not guaranteed to be
the first scene (or even present) in a multi-scene app

Fixes #1857

**Edit:** I originally reported that this would only fix the second
crash reported in #1857. Upon review of the first crash in that issue, I
believe this PR actually addresses both crashes.

## Test code and steps to reproduce
There really isn't any simple code that can be used to test this, as
this scenario requires having an app that implements multiple scenes, at
least one of which being CarPlay. It would be several hundred lines of
code.

## Checklist

- [x] Ensured that CI passes