fix(TVOS): fix crash with searchResultsController in RNSSearchBar (#2004)
## Description

On TVOS applications are crashing when user tries to open a screen with
Search Bar. This happens because TVOS requires UISearchController to
contain `searchResultsController`. this PR resolves it by adding simple
`UIViewController` as searchResultsController.

Fixes #2003.

## Changes

- Fixed crash on TVOS when user tries to open screen with search bar.

## Screenshots / GIFs

### Before


https://github.com/software-mansion/react-native-screens/assets/23281839/4f7b727a-139e-45a0-8141-f40b3e45d101

### After

> [!WARNING]
Please keep in mind that the visible search bar on that video does not
belong to the original implementation on TVOS.

https://github.com/software-mansion/react-native-screens/assets/23281839/7138f27e-294b-4d66-b5aa-ff2745d0d406

## Checklist

- [X] Ensured that CI passes