fix: Change postinstall step to prepare and run prepare on release in package.json (#2109)
## Description

This PR intents to fix error, during the install step on other React
Native applications. That's because of the name of the step - it
shouldn't be named as `postinstall`, as method named this way is
triggered by yarn v4. This was not an issue before, as we used yarn v1.

Fixes #2108.

## Changes

- Changed `postinstall` name to `prepare` in package.json
- Added `yarn prepare` command during the `yarn release`.

## Checklist

- [ ] Ensured that CI passes