chore(android): Implement missing codegen interfaces (#2163)
## Description

Currently when we change native props autogenerated by codegen we need
to manually update changed files. We're going to automate so everything
is copied automatically, but as a first step we need to implement those
that aren't yet. This is what this PR introduces.

## Test code and steps to reproduce

Go through Example app, nothing should change.