fix(iOS): non-interactive screen while switching between bottom-tab and native-stack navigators (#2260)
## Description
Currently, while switching between BottomTabNavigator and
NativeStackNavigator, user cannot interact with screens from
BottomTabNavigator. That's because while we're coming back to the
previous screen, original screen frame has frame (0, 0, 0, 0). I've
investigated that even if we're calling updateLayoutMetrics on a view,
it's not being updated, since the RNSScreen implementation of that
method firstly checks if `parentVC` is not RNSNavigationContainer and if
it's not nil -> Because it's always nil during the insertion (as we're
calling for screen updates too late), we're never updating layout
metrics.

Fixes #2252.

## Changes

- Moved calling for parent updates before updating layout metrics.

## Screenshots / GIFs

### Before


https://github.com/user-attachments/assets/8da01059-e488-43e4-ab44-2286b7cd3078

### After


https://github.com/user-attachments/assets/b7336425-bb01-4c66-80e7-4e8d83ca995d


## Test code and steps to reproduce

You can test `Test2252.tsx` to check these changes.


## Checklist

- [X] Included code example that can be used to test this change
- [x] Ensured that CI passes