fix(CI): extend logging in architecture-integrity scripts & add NativeProxy.kt to blacklist (#2281)
## Description

I was initially confused why the CI fails, so I debugged it a bit and it
turned out that we forgot to add `NativeProxy.kt` to blacklist in the
integrity scripts. <- Did it.

Because of the initial confusion I've decided to add some more logging &
error handling mechanisms.

## Test code and steps to reproduce

CI now passes. If similar case happens in the future we will get a nice
error message.

## Checklist

- [x] Ensured that CI passes