fix(Android,Fabric,bridgeless): crash on RN hot reload in dev mode when redbox in presentation (#2289)
## Description

There was a crash on Android + Fabric + bridgeless:

1. trigger a redbox by triggering some JS error (e.g. call a method that
does not exist),
2. fast refresh react-native,
3. see the crash in `onHostResume`.

The application crashed because we had a check for not null activity in
reactContext.
However after such reload ReactContext is recreated with null
activity!!!.

In such case, when we don't have access to the activity, we're currently
unable to create the dummy layout,
so the best we can do is not fail. The fix for jumping layout won't
work, however.

This PR also adds synchornization on `synchornied` block in
`maybeInitDummyLayoutWithHeaderMethod`, which should prevent
a possible data race, where both threads could try to create dummy
layout simmultaneously.

## Changes

* Do not fail in `onHostResume` when there is no activity present in the
provided context,
* only log a warning in the described case,
* synchronize layout initialisation, so that there is no data race <-
I'm not sure it was real due to JVM atomicity guarantees, however this
time I prefer the "better safe than sorry" approach,
* mark `isLayoutInitialized` variable as `Volatile`, to ensure that
updates from one thread are visible on the other thread.


## Test code and steps to reproduce

Desribed above ^ in the "Description" section.

## Checklist

- [ ] Ensured that CI passes